requirements.txt
Django==3.0.6
djangorestframework==3.12.2
django-treebeard==4.5.1
psycopg2-binary==2.8.6
